% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NsfR.R
\name{binaryScore}
\alias{binaryScore}
\title{run BinaryScore}
\usage{
binaryScore(negOutDF, informativeGenes, medianValues, column, nclust)
}
\arguments{
\item{negOutDF}{data.frame output from negativeOut()}

\item{informativeGenes}{numeric(1) number of genes to regard as informative for each cluster}

\item{medianValues}{data.frame cluster median expression values}

\item{column}{numeric(1) column number}

\item{nclust}{numeric(1) number of clusters}
}
\description{
run BinaryScore
}
