% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NsfR.R
\name{dtCutoff}
\alias{dtCutoff}
\title{to get expression cutoffs for f-beta testing}
\usage{
dtCutoff(binaryDF, clusterCol, dataDummy, dataFull)
}
\arguments{
\item{binaryDF}{data.frame}

\item{clusterCol}{numeric(1) cluster column}

\item{dataDummy}{matrix dummy columns for one vs all Random Forest modeling}

\item{dataFull}{data.frame tsvfile for from a SingleCellExperiment}
}
\description{
to get expression cutoffs for f-beta testing
}
