% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NsfR.R
\name{genPredDF}
\alias{genPredDF}
\title{to make the prediction columns from trees}
\usage{
genPredDF(cutoffs, dataFull)
}
\arguments{
\item{cutoffs}{data.frame holds cutoff values for each symbol}

\item{dataFull}{data.frame tsvfile for from a SingleCellExperiment}
}
\description{
to make the prediction columns from trees
}
