% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSForestR.R, R/NsfR.R
\name{negativeOut}
\alias{negativeOut}
\title{run negativeOut on output from randomForest()}
\usage{
negativeOut(impDF, column, medianValues, MedianExpressionLevel)

negativeOut(impDF, column, medianValues, MedianExpressionLevel)
}
\arguments{
\item{impDF}{data.frame}

\item{column}{numeric(1) column number}

\item{medianValues}{data.frame  cluster median expression values}

\item{MedianExpressionLevel}{numeric(1) parameter for filtering and ranking genes}

\item{impDF}{data.frame}

\item{column}{numeric(1) column number}

\item{medianValues}{data.frame  cluster median expression values}

\item{MedianExpressionLevel}{numeric(1) parameter for filtering and ranking genes}
}
\description{
run negativeOut on output from randomForest()

run negativeOut on output from randomForest()
}
