% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSForestR.R
\name{NSForestCL}
\alias{NSForestCL}
\title{run NSForestCL}
\usage{
NSForestCL(file, clusterColumn, clusterValue, rfTrees, informativeGenes,
  num_top_ranked, betaValue, column)
}
\arguments{
\item{file}{csv file}

\item{clusterColumn}{numeric(1) cluster column}

\item{rfTrees}{numeric(1) number of trees}

\item{informativeGenes}{numeric(1) number of genes to regard as informative for each cluster}

\item{num_top_ranked}{numeric(1) number of symbols to retain}

\item{betaValue}{numeric(1) value for beta weighting}

\item{column}{numeric(1) column number}

\item{cluster_value}{numeric(1) value for cluster}

\item{medianValues}{data.frame  cluster median expression values}
}
\description{
run NSForestCL
}
