% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNSFromPython.R
\name{runNSFromPython}
\alias{runNSFromPython}
\title{to run NS Forest program from python}
\usage{
runNSFromPython(sce, numtrees = 100, numthread = 1,
  expressionLevel = 0, num_inf_genes = 5, num_top_ranked = 3,
  beta_value = 0.5)
}
\arguments{
\item{sce}{instance of SingleCellExperiment}

\item{numtrees}{numeric(1) number of trees built in random forest step}

\item{numthread}{numeric(1) number of threads for python execution}

\item{expressionLevel}{numeric(1) median expression level}

\item{num_inf_genes}{numeric(1) number of genes to regard as informative for each cluster}

\item{num_top_ranked}{numeric(1) number of genes to retain in reporting}

\item{beta_value}{numeric(1) set values for fbeta weighting}
}
\description{
to run NS Forest program from python
}
\examples{
require("pcmp")
data(Ab10k)
r1 = runNSFromPython(Ab10k) # running with default parameters
}
