% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NsfR.R
\name{runRandomForest}
\alias{runRandomForest}
\title{run RandomForest}
\usage{
runRandomForest(column, dataFull, dataDummy, rfTrees, cind = cind)
}
\arguments{
\item{column}{numeric(1) column number}

\item{dataFull}{data.frame tsvfile for from a SingleCellExperiment}

\item{dataDummy}{matrix dummy columns for one vs all Random Forest modeling}

\item{rfTrees}{numeric(1) number of trees}

\item{cind}{integer index of "Clusters" column}
}
\description{
run RandomForest
}
