% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceTotsv.R
\name{sceTotsv}
\alias{sceTotsv}
\title{transform sce to tsv file}
\usage{
sceTotsv(sce, destination)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{destination}{character filepath to tsv}
}
\description{
transform sce to tsv file
}
\examples{
require(pcmp)
data(Ab10k)
tf = tempfile(fileext=".tsv")
sceTotsv(Ab10k,tf)
head(read.delim(tf,sep="\\t"))
}
