% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSForestR.R, R/NsfR.R
\name{fBetaTest}
\alias{fBetaTest}
\title{run fBeta test on prediction results}
\usage{
fBetaTest(predDF, clusterNum, betaValue, maxSymbol)

fBetaTest(predDF, clusterNum, betaValue, maxSymbol)
}
\arguments{
\item{predDF}{data.frame prediction results for each symbol}

\item{betaValue}{numeric(1) value for beta weighting}

\item{maxSymbol}{numeric(1) number of symbols to retain}

\item{clusterName}{numeric(1) cluster number}

\item{predDF}{data.frame prediction results for each symbol}

\item{clusterName}{numeric(1) cluster number}

\item{betaValue}{numeric(1) value for beta weighting}

\item{maxSymbol}{numeric(1) number of symbols to retain}
}
\description{
run fBeta test on prediction results

run fBeta test on prediction results
}
