% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSForestR.R
\name{fBetaTest}
\alias{fBetaTest}
\title{run fBeta test on prediction results}
\usage{
fBetaTest(predDF, clusterNum, betaValue, maxSymbol, dataDummy)
}
\arguments{
\item{predDF}{data.frame prediction results for each symbol}

\item{betaValue}{numeric(1) value for beta weighting}

\item{maxSymbol}{numeric(1) number of symbols to retain}

\item{dataDummy}{matrix dummy columns for one vs all Random Forest modeling}

\item{clusterName}{numeric(1) cluster number}
}
\description{
run fBeta test on prediction results
}
