% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subst.R
\name{runNSForest}
\alias{runNSForest}
\title{run doSubs to generate python program and source
from reticulate}
\usage{
runNSForest()
}
\description{
run doSubs to generate python program and source
from reticulate
}
