% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNSFromPython.R
\name{runNSFromPython}
\alias{runNSFromPython}
\title{to run NS Forest program from python}
\usage{
runNSFromPython(tsvfile, rfTrees, threads, Median_Expression_Level,
  InformativeGenes, Genes_to_testing, betaValue)
}
\arguments{
\item{tsvfile}{}

\item{rfTrees}{numeric(1) number of trees built in random forest step}

\item{threads}{numeric(1) number of threads for python execution}

\item{Median_Expression_Level}{numeric(1) median expression level}

\item{InformativeGenes}{numeric(1) number of genes to regard as informative for each cluster}

\item{Genes_to_testing}{numeric(1) number of genes to retain in reporting}

\item{betaValue}{numeric(1) set values for fbeta weighting}
}
\description{
to run NS Forest program from python
}
\examples{
tsvfile = system.file("data/Ab10k.tsv", package="NSForestR")
runNSFromPython(tsvfile, rfTrees=100L, threads=1L, Median_Expression_Level=0L, InformativeGenes=5L,
                Genes_to_testing=3L, betaValue=0.5)
}
