% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceTotsv.R
\name{sceTotsv}
\alias{sceTotsv}
\title{transform sce to tsv file}
\usage{
sceTotsv(sce, destination)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\description{
transform sce to tsv file
}
\examples{
require(pcmp)
data(Ab10k)
sceTotsv(Ab10k,"temp/Ab10k.tsv")
}
